+++ src/components/Label/Label.stories.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+import Label from "./Label";
+
+// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
+export default {
+  title: "Atoms/Label",
+  component: Label,
+  argTypes: { onClick: { action: "clicked" } },
+} as ComponentMeta<typeof Label>;
+
+// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
+const Template: ComponentStory<typeof Label> = (args) => <Label {...args} />;
+
+export const Basic = Template.bind({});
+// More on args: https://storybook.js.org/docs/react/writing-stories/args
+Basic.args = {
+  for: "text-input",
+  children: "Label",
+};
+Basic.parameters = {
+  docs: {
+    source: {
+      code: '<Label for="example-input">Label</Label>',
+    },
+  },
+};