+++ src/components/index.ts
@@ -1,6 +1,8 @@
 // atoms
 export { default as Button } from "./Button";
 export { default as Input } from "./Input";
+export { default as Label } from "./Label";
+export { default as LabeledInput } from "./LabeledInput";
 // patterns
 export { default as Layout } from "./Layout";
 export { default as Navigation } from "./Navigation";